// src/app/shared/common/apis/generated/content/model/originalAsset.ts
/**
 * Goddard.ContentWebApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Entity } from './entity';
import { Action } from './action';
import { Asset } from './asset';
import { Link } from './link';


export interface OriginalAsset { 
    _class?: Array<string> | null;
    properties?: { [key: string]: any; } | null;
    entities?: Array<Entity> | null;
    links?: Array<Link> | null;
    actions?: Array<Action> | null;
    title?: string | null;
    rel?: Array<string> | null;
    href?: string | null;
    type?: string | null;
    readonly name?: string | null;
    readonly contentPath?: string | null;
    publishUrl?: string | null;
    readonly url?: string | null;
    renditions?: Array<Asset> | null;
}